Implement special backoff behavior on HTTP 409
It's clear from the documentation that Google may respond with a message
asking the client to back off, but it's not clear what form that takes
over the HTTP API. It seems though it could be an HTTP 409, so make an
effort to parse and respect that.

Since the looping and retry logic was getting complicated enough, it's
been extracted to a separate Looper class. Besides the new behavior on
HTTP 4xx responses, it should be unchanged.